Assets now must get an ID at creation time. Added a possibly naive id generator

+++ engine/include/EngineTypedefs.hpp
@@ -9,10 +9,12 @@ typedef uint32_t Index;
 typedef uint32_t Offset;
 typedef uint32_t WindowSize;
 typedef uint32_t WindowHandle;
-typedef uint32_t GameObjectID;
-typedef uint32_t ComponentID;
-typedef uint32_t AssetID;
+typedef uint64_t GameObjectID;
+typedef uint64_t ComponentID;
+typedef uint64_t AssetID;
 typedef uint64_t ElementCount;
 typedef uint64_t Capacity;
 typedef uint64_t SizeBytes;
+
+
 } // namespace Engine
